+++ src/Fable.Transforms/Python/Fable2Python.fs
@@ -1118,8 +1118,8 @@ module Util =
 
             match memberInfo.CurriedParameterGroups, memberInfo.DeclaringEntity with
             // Check only members with multiple non-curried arguments
-            | [parameters], Some ent when not (List.isEmpty parameters) ->
-                com.TryGetNonCoreAssemblyEntity(ent)
+            | [parameters], Some entRef when not (List.isEmpty parameters) ->
+                com.TryGetEntity(entRef)
                 |> Option.bind (fun ent ->
                     if ent.IsFSharpModule then None
                     else
@@ -2733,10 +2733,9 @@ module Compiler =
             member _.OutputDir = com.OutputDir
             member _.OutputType = com.OutputType
             member _.ProjectFile = com.ProjectFile
-            member _.GetEntity(fullName) = com.GetEntity(fullName)
-            member _.TryGetNonCoreAssemblyEntity(fullName) = com.TryGetNonCoreAssemblyEntity(fullName)
             member _.GetImplementationFile(fileName) = com.GetImplementationFile(fileName)
             member _.GetRootModule(fileName) = com.GetRootModule(fileName)
+            member _.TryGetEntity(fullName) = com.TryGetEntity(fullName)
             member _.GetInlineExpr(fullName) = com.GetInlineExpr(fullName)
             member _.AddWatchDependency(fileName) = com.AddWatchDependency(fileName)
             member _.AddLog(msg, severity, ?range, ?fileName:string, ?tag: string) =