+++ src/Fable.Cli/RELEASE_NOTES.md
@@ -1,3 +1,14 @@
+### 3.6.3
+
+* New FSharp.Core 6 APIs: updateAt/insertAt/insertManyAt/removeAt/removeManyAt
+* Support thousand separators in String.Format
+* Fix #2628 @stroborobo
+
+### 3.6.2
+
+* TypescriptTaggedUnion @cannorin
+* Speed up recompilation when adding/removing files
+
 ### 3.6.1
 
 * Fix #2614: Char addition